requirements.txt
Django
django-tinymce4-lite
